requirements.txt
Flask
requests
beautifulsoup4
matplotlib
python-dotenv

